requirements_backend.txt
google-cloud-storage
python-multipart
torch
torchvision
timm
fastapi
uvicorn
gsutil
requests
google-cloud-run
hydra-core